+++ src/components/list/List.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { List, Typography } from 'antd';
+
+
+class List extends React.Component {
+    render() {
+
+    }   
+
+}