+++ src/pages/terms/Terms.js
@@ -1,13 +1,16 @@
 import React, { PureComponent } from 'react';
+import TermsStore from './store/TermsStore';
+import TermsChild from './TermsChild';
 
 // @flow
-// 여기서 mobx를 이용하여 api 테스트 할 것
-
 class Terms extends PureComponent {
     render() {
         return (
-            <div>
-                <h1>terms</h1>
+            <div class="dom-parent">
+                <h1>terms</h1>                
+                <div class="termsArea">
+                    <TermsChild store={TermsStore} />
+                </div>
             </div>
         )
     }